Use `link-to` instead of using a `transition-to` helper for buttons and move actions from route to controller.
