+++ app/components/article-meta.js
@@ -3,13 +3,24 @@ import { inject as service } from '@ember/service';
 
 export default Component.extend({
   session: service(),
+
+  /**
+   * Flag to help determine whether to render a logged-in user state or not
+   * @type {Boolean}
+   */
+  isLoggedIn: false,
+
   /**
    * Method executed when "favorite" button is clicked.
+   * @method
+   * @name onFavoriteArticle
+   * @param {Object} article The article record to be favorited
    */
-  onFavoriteArticle() {},
 
   /**
    * Method executed when "follow" button is clicked.
+   * @method
+   * @name onFollowAuthor
+   * @param {Object} author The author record that is followed
    */
-  onFollowAuthor() {},
 });
