+++ app/templates/components/article-meta.hbs
@@ -1,13 +1,24 @@
 <div class="article-meta" data-test-article-meta>
   {{article-author article=article tagName=""}}
-  <button class="btn btn-sm {{if article.author.following "btn-secondary" "btn-outline-secondary"}}"
-    data-test-follow-author-button {{action onFollowAuthor}}>
-    <i class="ion-plus-round"></i> &nbsp; {{if article.author.following "Unfollow" "Follow"}} {{article.author.id}}
-  </button>
-  &nbsp;&nbsp;
-  <button class="btn btn-sm {{if article.favorited "btn-primary" "btn-outline-primary"}}" {{action onFavoriteArticle}}
-    data-test-favorite-article-button>
-    <i class="ion-heart"></i> &nbsp; {{if article.favorited "Unfavorite" "Favorite"}} Post
-    <span class="counter" data-test-article-favorites-count>{{article.favoritesCount}}</span>
-  </button>
+  {{#if isLoggedIn}}
+    <button class="btn btn-sm {{if article.author.following "btn-secondary" "btn-outline-secondary"}}"
+      data-test-follow-author-button {{action onFollowAuthor article.author}}>
+      <i class="ion-plus-round"></i> &nbsp; {{if article.author.following "Unfollow" "Follow"}} {{article.author.id}}
+    </button>
+    &nbsp;&nbsp;
+    <button class="btn btn-sm {{if article.favorited "btn-primary" "btn-outline-primary"}}"
+      {{action onFavoriteArticle article}} data-test-favorite-article-button>
+      <i class="ion-heart"></i> &nbsp; {{if article.favorited "Unfavorite" "Favorite"}} Post
+      <span class="counter" data-test-article-favorites-count>{{article.favoritesCount}}</span>
+    </button>
+  {{else}}
+    {{#link-to "login" tagName="button" class="btn btn-sm btn-outline-secondary" data-test-follow-author-button-login=""}}
+      Follow {{article.author.id}}
+    {{/link-to}}
+    &nbsp;&nbsp;
+    {{#link-to "login" tagName="button" class="btn btn-sm btn-outline-secondary" data-test-favorite-article-button-login=""}}
+      <i class="ion-heart"></i> &nbsp; Favorite Post
+      <span class="counter" data-test-article-favorites-count>{{article.favoritesCount}}</span>
+    {{/link-to}}
+  {{/if}}
 </div>
